% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_selection.R
\name{analyze_pivots}
\alias{analyze_pivots}
\title{Analyze the pivot selection behavior of rpcholesky}
\usage{
analyze_pivots(A, result, compute_eigendecomp = TRUE, plot = TRUE)
}
\arguments{
\item{A}{Matrix being approximated}

\item{result}{Result from rpcholesky}

\item{compute_eigendecomp}{Whether to compute eigendecomposition}

\item{plot}{Whether to create visualizations}
}
\value{
Analysis of pivot selection
}
\description{
Analyze the pivot selection behavior of rpcholesky
}
