% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{kernel_alignment}
\alias{kernel_alignment}
\title{Calculate kernel alignment between original and approximated matrices}
\usage{
kernel_alignment(K, K_approx, factor = TRUE)
}
\arguments{
\item{K}{Original kernel matrix}

\item{K_approx}{Approximated kernel matrix or approximation components}

\item{factor}{Whether K_approx contains the factors (G) or the complete approximation}
}
\value{
A scalar representing the alignment between the matrices (1 is perfect)
}
\description{
Calculate kernel alignment between original and approximated matrices
}
