% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpcholesky-package.R
\name{rpcholesky}
\alias{rpcholesky}
\title{Accelerated Randomly Pivoted Cholesky Decomposition}
\usage{
rpcholesky(A, k, b = "auto", stoptol = 1e-10, verbose = FALSE)
}
\arguments{
\item{A}{A positive semidefinite matrix}

\item{k}{Target rank for the approximation}

\item{b}{Batch size (an integer or "auto")}

\item{stoptol}{Early stopping tolerance}

\item{verbose}{If TRUE, prints progress information}
}
\value{
A list containing low-rank factor G, selected pivot indices, and corresponding rows
}
\description{
Accelerated Randomly Pivoted Cholesky Decomposition
}
