% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_metrics.R
\name{rpcholesky_error}
\alias{rpcholesky_error}
\title{Calculate comprehensive error metrics for low-rank approximation}
\usage{
rpcholesky_error(A, approx, factor = TRUE)
}
\arguments{
\item{A}{Original positive semidefinite matrix}

\item{approx}{Approximated matrix or approximation components}

\item{factor}{Whether approx contains the factors (G) or the complete approximation (G^T * G)}
}
\value{
A named list of various error metrics
}
\description{
Calculate comprehensive error metrics for low-rank approximation
}
